+++ docs/src/content/ja/docs/auth/_meta.ts
@@ -3,5 +3,6 @@ const meta = {
   jwt: "JSON Web Token",
   clerk: "Clerk",
   supabase: "Supabase",
+  firebase: "Firebase",
 };
 export default meta;